requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==70.3.0
sqlparse==0.5.0
uritemplate==4.1.1
